review end points and task models
