--- app/src/main/java/com/homies/homies/retrofit/model/tasksModels/CreateNewTaskResponse.java
@@ -1,73 +1,0 @@
-package com.homies.homies.retrofit.model.tasksModels;
-
-import android.service.autofill.UserData;
-
-import com.homies.homies.retrofit.model.TaskList;
-import com.homies.homies.retrofit.model.UserAdmin;
-
-public class CreateNewTaskResponse {
-
-    private int id;
-    private int groupKey;
-    private String groupName;
-    private String groupRelation;
-    private UserData userData;
-    private UserAdmin userAdmin;
-    private TaskList taskList;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public int getGroupKey() {
-        return groupKey;
-    }
-
-    public void setGroupKey(int groupKey) {
-        this.groupKey = groupKey;
-    }
-
-    public String getGroupName() {
-        return groupName;
-    }
-
-    public void setGroupName(String groupName) {
-        this.groupName = groupName;
-    }
-
-    public String getGroupRelation() {
-        return groupRelation;
-    }
-
-    public void setGroupRelation(String groupRelation) {
-        this.groupRelation = groupRelation;
-    }
-
-    public UserData getUserData() {
-        return userData;
-    }
-
-    public void setUserData(UserData userData) {
-        this.userData = userData;
-    }
-
-    public UserAdmin getUserAdmin() {
-        return userAdmin;
-    }
-
-    public void setUserAdmin(UserAdmin userAdmin) {
-        this.userAdmin = userAdmin;
-    }
-
-    public TaskList getTaskList() {
-        return taskList;
-    }
-
-    public void setTaskList(TaskList taskList) {
-        this.taskList = taskList;
-    }
-}
