--- app/src/main/java/com/homies/homies/retrofit/model/tasksModels/UpdateTaskResponse.java
@@ -1,115 +1,0 @@
-package com.homies.homies.retrofit.model.tasksModels;
-
-import com.homies.homies.retrofit.model.TaskList;
-import com.homies.homies.retrofit.model.UserData;
-
-public class UpdateTaskResponse {
-    private int id;
-    private String taskName;
-    private String dataCreate;
-    private String dataEnd;
-    private String description;
-    private boolean cancel;
-    private String photo;
-    private String photoContentType;
-    private int puntuacion;
-    private TaskList taskList;
-    private UserData userData;
-    private String userAssigneds;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getTaskName() {
-        return taskName;
-    }
-
-    public void setTaskName(String taskName) {
-        this.taskName = taskName;
-    }
-
-    public String getDataCreate() {
-        return dataCreate;
-    }
-
-    public void setDataCreate(String dataCreate) {
-        this.dataCreate = dataCreate;
-    }
-
-    public String getDataEnd() {
-        return dataEnd;
-    }
-
-    public void setDataEnd(String dataEnd) {
-        this.dataEnd = dataEnd;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public boolean isCancel() {
-        return cancel;
-    }
-
-    public void setCancel(boolean cancel) {
-        this.cancel = cancel;
-    }
-
-    public String getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(String photo) {
-        this.photo = photo;
-    }
-
-    public String getPhotoContentType() {
-        return photoContentType;
-    }
-
-    public void setPhotoContentType(String photoContentType) {
-        this.photoContentType = photoContentType;
-    }
-
-    public int getPuntuacion() {
-        return puntuacion;
-    }
-
-    public void setPuntuacion(int puntuacion) {
-        this.puntuacion = puntuacion;
-    }
-
-    public TaskList getTaskList() {
-        return taskList;
-    }
-
-    public void setTaskList(TaskList taskList) {
-        this.taskList = taskList;
-    }
-
-    public UserData getUserData() {
-        return userData;
-    }
-
-    public void setUserData(UserData userData) {
-        this.userData = userData;
-    }
-
-    public String getUserAssigneds() {
-        return userAssigneds;
-    }
-
-    public void setUserAssigneds(String userAssigneds) {
-        this.userAssigneds = userAssigneds;
-    }
-}
